Remove StartupBase
Remove StartupBase and just have a Startup class.
