+++ src/Website/Startup.cs
@@ -3,13 +3,27 @@
 
 namespace MartinCostello.Website
 {
+    using System;
+    using Extensions;
+    using Microsoft.AspNetCore.Builder;
+    using Microsoft.AspNetCore.CookiePolicy;
     using Microsoft.AspNetCore.Hosting;
+    using Microsoft.AspNetCore.Http;
+    using Microsoft.AspNetCore.HttpOverrides;
+    using Microsoft.AspNetCore.Mvc;
     using Microsoft.Extensions.Configuration;
+    using Microsoft.Extensions.DependencyInjection;
+    using Microsoft.Extensions.Options;
+    using Microsoft.Net.Http.Headers;
+    using Newtonsoft.Json;
+    using NodaTime;
+    using Options;
+    using Services;
 
     /// <summary>
     /// A class representing the startup logic for the application.
     /// </summary>
-    public class Startup : StartupBase
+    public class Startup
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="Startup"/> class.
@@ -17,8 +31,188 @@ namespace MartinCostello.Website
         /// <param name="configuration">The <see cref="IConfiguration"/> to use.</param>
         /// <param name="hostingEnvironment">The <see cref="IHostingEnvironment"/> to use.</param>
         public Startup(IConfiguration configuration, IHostingEnvironment hostingEnvironment)
-            : base(configuration, hostingEnvironment)
         {
+            Configuration = configuration;
+            HostingEnvironment = hostingEnvironment;
+        }
+
+        /// <summary>
+        /// Gets the current configuration.
+        /// </summary>
+        public IConfiguration Configuration { get; }
+
+        /// <summary>
+        /// Gets the current hosting environment.
+        /// </summary>
+        public IHostingEnvironment HostingEnvironment { get; }
+
+        /// <summary>
+        /// Gets or sets the service provider.
+        /// </summary>
+        public IServiceProvider ServiceProvider { get; set; }
+
+        /// <summary>
+        /// Configures the application.
+        /// </summary>
+        /// <param name="app">The <see cref="IApplicationBuilder"/> to use.</param>
+        /// <param name="serviceProvider">The <see cref="IServiceProvider"/> to use.</param>
+        /// <param name="options">The snapshot of <see cref="SiteOptions"/> to use.</param>
+        public void Configure(IApplicationBuilder app, IServiceProvider serviceProvider, IOptionsSnapshot<SiteOptions> options)
+        {
+            ServiceProvider = serviceProvider;
+
+            app.UseCustomHttpHeaders(HostingEnvironment, Configuration, options);
+
+            if (HostingEnvironment.IsDevelopment())
+            {
+                app.UseDeveloperExceptionPage();
+            }
+            else
+            {
+                app.UseExceptionHandler("/error")
+                   .UseStatusCodePagesWithReExecute("/error", "?id={0}");
+            }
+
+            app.UseHsts()
+               .UseHttpsRedirection();
+
+            app.UseStaticFiles(
+                new StaticFileOptions()
+                {
+                    DefaultContentType = "application/json",
+                    ServeUnknownFileTypes = true,
+                    OnPrepareResponse = (context) =>
+                        {
+                            var headers = context.Context.Response.GetTypedHeaders();
+                            headers.CacheControl = new CacheControlHeaderValue()
+                            {
+                                MaxAge = TimeSpan.FromDays(7)
+                            };
+                        }
+                });
+
+            app.UseForwardedHeaders(
+                new ForwardedHeadersOptions()
+                {
+                    ForwardedHeaders = ForwardedHeaders.XForwardedFor | ForwardedHeaders.XForwardedProto
+                });
+
+            app.UseHttpMethodOverride();
+
+            app.UseMvcWithDefaultRoute();
+
+            app.UseCookiePolicy(CreateCookiePolicy());
+        }
+
+        /// <summary>
+        /// Configures the services for the application.
+        /// </summary>
+        /// <param name="services">The <see cref="IServiceCollection"/> to use.</param>
+        public void ConfigureServices(IServiceCollection services)
+        {
+            services.AddApplicationInsightsTelemetry(Configuration);
+            services.AddOptions();
+            services.Configure<SiteOptions>(Configuration.GetSection("Site"));
+
+            services.AddAntiforgery(
+                (p) =>
+                {
+                    p.Cookie.HttpOnly = true;
+                    p.Cookie.Name = "_anti-forgery";
+                    p.Cookie.SecurePolicy = CookiePolicy();
+                    p.FormFieldName = "_anti-forgery";
+                    p.HeaderName = "x-anti-forgery";
+                });
+
+            services
+                .AddMemoryCache()
+                .AddDistributedMemoryCache()
+                .AddMvc(ConfigureMvc)
+                .AddJsonOptions((p) => services.AddSingleton(ConfigureJsonFormatter(p)));
+
+            services.AddRouting(
+                (p) =>
+                {
+                    p.AppendTrailingSlash = true;
+                    p.LowercaseUrls = true;
+                });
+
+            services.AddHsts(
+                (p) =>
+                {
+                    p.MaxAge = TimeSpan.FromDays(365);
+                    p.IncludeSubDomains = false;
+                    p.Preload = false;
+                });
+
+            services
+                .AddResponseCaching()
+                .AddResponseCompression();
+
+            services.AddSingleton<IClock>((_) => SystemClock.Instance);
+
+            services.AddScoped((p) => p.GetRequiredService<IHttpContextAccessor>().HttpContext);
+            services.AddScoped((p) => p.GetRequiredService<IOptionsSnapshot<SiteOptions>>().Value);
+            services.AddScoped<IToolsService, ToolsService>();
+        }
+
+        /// <summary>
+        /// Configures the JSON serializer for MVC.
+        /// </summary>
+        /// <param name="options">The <see cref="MvcJsonOptions"/> to configure.</param>
+        /// <returns>
+        /// The <see cref="JsonSerializerSettings"/> to use.
+        /// </returns>
+        private static JsonSerializerSettings ConfigureJsonFormatter(MvcJsonOptions options)
+        {
+            // Make JSON easier to read for debugging at the expense of larger payloads
+            options.SerializerSettings.Formatting = Formatting.Indented;
+
+            // Omit nulls to reduce payload size
+            options.SerializerSettings.NullValueHandling = NullValueHandling.Ignore;
+
+            // Explicitly define behavior when serializing DateTime values
+            options.SerializerSettings.DateFormatString = "yyyy'-'MM'-'dd'T'HH':'mm':'ssK";   // Only return DateTimes to a 1 second precision
+
+            return options.SerializerSettings;
+        }
+
+        /// <summary>
+        /// Configures MVC.
+        /// </summary>
+        /// <param name="options">The <see cref="MvcOptions"/> to configure.</param>
+        private void ConfigureMvc(MvcOptions options)
+        {
+            if (!HostingEnvironment.IsDevelopment())
+            {
+                options.Filters.Add(new RequireHttpsAttribute());
+            }
+        }
+
+        /// <summary>
+        /// Creates the <see cref="CookiePolicyOptions"/> to use.
+        /// </summary>
+        /// <returns>
+        /// The <see cref="CookiePolicyOptions"/> to use for the application.
+        /// </returns>
+        private CookiePolicyOptions CreateCookiePolicy()
+        {
+            return new CookiePolicyOptions()
+            {
+                HttpOnly = HttpOnlyPolicy.Always,
+                Secure = CookiePolicy(),
+            };
+        }
+
+        /// <summary>
+        /// Creates the <see cref="CookieSecurePolicy"/> to use.
+        /// </summary>
+        /// <returns>
+        /// The <see cref="CookieSecurePolicy"/> to use for the application.
+        /// </returns>
+        private CookieSecurePolicy CookiePolicy()
+        {
+            return HostingEnvironment.IsDevelopment() ? CookieSecurePolicy.SameAsRequest : CookieSecurePolicy.Always;
         }
     }
 }
